requirements.txt
# Dependencies for Packit AI workflows
ogr
requests
pyyaml
